api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Jaeyoung Park <alska1039@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
mysql-connector-python = "^8.4.0"
scikit-learn = "^1.4.2"

konlpy = "^0.6.0"
gensim = "^4.3.2"
scipy = "1.10.1"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.0"
matplotlib = "^3.9.0"
koreanize-matplotlib = "^0.1.1"
pandas = "^2.2.2"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
